Funcionalidad completa

+++ WebApiSegura/Models/Factura.cs
@@ -10,9 +10,9 @@ namespace WebApiSegura.Models
         public int FACTURA_ID { get; set; }
         public int USUARIO_ID { get; set; }
         public int PLAN_ID { get; set; }
-        public double MONTO_FACTURA { get; set; }
+        public decimal MONTO_FACTURA { get; set; }
         public int CANT_PRODUCTOS { get; set; }
         public string ESTADO { get; set; }
-        public double PAGO_MENSUAL { get; set; }
+        public decimal PAGO_MENSUAL { get; set; }
     }
 }
